# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
  model.User:
    properties:
      age:
        example: 25
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Hiro
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 練習用 Go lang 測試
  title: GoLang API Practice
  version: 1.0.0
paths:
  /test/hello/{name}:
    get:
      description: 回應 Hello Name
      parameters:
      - description: 使用者名稱
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hello
      tags:
      - Test
  /test/ping:
    get:
      description: Health check
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping
      tags:
      - Test
  /user/{id}:
    get:
      description: 根據 ID 取得使用者
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get Single User
      tags:
      - User
  /user/all:
    get:
      description: 取得所有使用者
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get All Users
      tags:
      - User
swagger: "2.0"